+++ entrypoint.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-# VERSION: 0.0.27
+# VERSION: 0.0.28
 # AUTHORS: Ogekuri