+++ test_sviluppo_plugins.py
@@ -1,4 +1,4 @@
-# VERSION: 0.0.27
+# VERSION: 0.0.28
 # AUTHORS: Ogekuri
 
 from bs4 import BeautifulSoup