+++ utils/parse_config.py
@@ -1,35 +1,30 @@
-# VERSION: 0.0.27
+# VERSION: 0.0.28
 # AUTHORS: aymene69
 # CONTRIBUTORS: Ogekuri
 
-import json
-from lzstring import LZString
+from utils.string_encoding import decode_lzstring, encode_lzstring
 
-# from utils.string_encoding import decodeb64
-
-# def parse_config(b64config):
-#     config = json.loads(decodeb64(b64config))
-
-#     # For backwards compatibility
-#     if "languages" not in config:
-#         config["languages"] = [config["language"]]
-
-#     return config
-
+# wrapping alla decode_lzstring per gestire eventuali retro-compatibità
 def parse_config(encoded_config):
-    
-    lz = LZString()
-
-    # Se il prefisso "C_" è presente, rimuovilo
-    if encoded_config.startswith("C_"):
-        encoded_config = encoded_config[2:]
-    
-    # decodifica e decomprime utilizzando l'algoritmo di LZString con encodeURIComponent
-    decompressed = lz.decompressFromEncodedURIComponent(encoded_config)
-    config = json.loads(decompressed)
-
-    # Compatibilità retroattiva
-    if "languages" not in config and "language" in config:
-        config["languages"] = [config["language"]]
+        
+    # decodifica utilizzando l'algoritmo di LZString con encodeURIComponent
+    config = decode_lzstring(encoded_config, "C_")
 
     return config
+
+
+# wrapping alla decode_lzstring per gestire eventuali retro-compatibità
+def parse_query(encoded_query):
+        
+    # decodifica utilizzando l'algoritmo di LZString con encodeURIComponent
+    query = decode_lzstring(encoded_query, "Q_")
+
+    return query
+
+# wrapping alla encode_lzstring per gestire eventuali retro-compatibità
+def encode_query(query):
+        
+    # decodifica utilizzando l'algoritmo di LZString con encodeURIComponent
+    encoded_query = encode_lzstring(query, "Q_")
+
+    return encoded_query