+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Elenco modifiche introdotti in Debriddo.
 
 Il formato è basato su on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) ed aderisce a [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.0.28](https://github.com/Ogekuri/debriddo/compare/v0.0.27...v0.0.28) (2024-12-12)
+
+- Modifica raise nella libreria Debrid
+- Ricodificata la query con LZ-String
+  
 ## [v0.0.27](https://github.com/Ogekuri/debriddo/compare/v0.0.26...v0.0.27) (2024-12-11)
 
 - Bu fix coroutine 'RealDebrid.delete_torrent' was never awaited